+++ converter/convert_test.go
@@ -14,6 +14,7 @@ func TestConvertFromJsonToTerraformHcl(t *testing.T) {
 		{"fixtures/simple.json", "fixtures/simple.tf"},
 		{"fixtures/empty-policy.json", "fixtures/empty-policy.tf"},
 		{"fixtures/single-statement.json", "fixtures/single-statement.tf"},
+		{"fixtures/tf-interpolations.json", "fixtures/tf-interpolations.tf"},
 	}
 
 	for _, fixture := range fixtures {
@@ -40,4 +41,40 @@ func TestConvertFromJsonToTerraformHcl(t *testing.T) {
 		})
 	}
 
+}
+
+func TestEscapeDollarSigns(t *testing.T) {
+	var testcases = []struct {
+		in  string
+		out string
+	}{
+		// basic strings
+		{"", ""},
+		{"hello world", "hello world"},
+		// aws policy variables
+		{"${aws:username}", "$${aws:username}"},
+		{"home/${aws:username}/*", "home/$${aws:username}/*"},
+		{"arn:aws:s3:::bucket/${aws:PrincipalTag/department}", "arn:aws:s3:::bucket/$${aws:PrincipalTag/department}"},
+		{"${s3:x-amz-acl}", "$${s3:x-amz-acl}"},
+
+		// aws special character escaping
+		{"${*}", "$${*}"},
+		{"${?}", "$${?}"},
+		{"${$}", "$${$}"},
+
+		// terraform interpolations
+		{"${aws_kms_key.key.arn}", "${aws_kms_key.key.arn}"},
+		{"hello ${var.greeting}!", "hello ${var.greeting}!"},
+		{"arn:aws:s3:::foo/${join(var.separator,local.path_elements)}", "arn:aws:s3:::foo/${join(var.separator,local.path_elements)}"},
+	}
+
+	for _, testcase := range testcases {
+		t.Run(testcase.in, func(t *testing.T) {
+			assert.EqualValues(t,
+				testcase.out,
+				escapePolicyVariables(testcase.in),
+			)
+		})
+
+	}
 }