+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This changelog contains a loose collection of changes in every release. I will a
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## Unreleased
+
+### Fixed
+
+* Limit escaping of dollar signs to IAM policy variables, don't escape terraform interpolations
+
 ## 1.4.0 - 2020-04-19
 
 ### Fixed